# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Postal Registration API",
        "description": "Postal Registration - Laravel API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/postal-packages": {
            "get": {
                "tags": [
                    "Postal Packages"
                ],
                "summary": "List all postal-packages",
                "operationId": "e3339a7df6604f8d48bbd3173326e427",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostalPackageResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Postal Packages"
                ],
                "summary": "Create a new postal package",
                "operationId": "2d8913398e420dab45c64dd047235aea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sender",
                                    "receiver",
                                    "weight",
                                    "dimensions"
                                ],
                                "properties": {
                                    "sender": {
                                        "required": [
                                            "name",
                                            "mobile",
                                            "postal_code",
                                            "address"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "علی رضایی"
                                            },
                                            "mobile": {
                                                "type": "string",
                                                "example": "09123456789"
                                            },
                                            "postal_code": {
                                                "type": "string",
                                                "example": "1234567890"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "تهران، خیابان ولیعصر"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "receiver": {
                                        "required": [
                                            "name",
                                            "mobile",
                                            "postal_code",
                                            "address"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "محمد حسینی"
                                            },
                                            "mobile": {
                                                "type": "string",
                                                "example": "09351234567"
                                            },
                                            "postal_code": {
                                                "type": "string",
                                                "example": "0987654321"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "مشهد، بلوار وکیل آباد"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "dimensions": {
                                        "required": [
                                            "length",
                                            "width",
                                            "height"
                                        ],
                                        "properties": {
                                            "length": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 30
                                            },
                                            "width": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 20
                                            },
                                            "height": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 15
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parcel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parcel created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tracking_code": {
                                                    "type": "string",
                                                    "example": "TRK170420234567"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-01-01 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "properties": {
                                                "sender.mobile": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "فرمت شماره موبایل فرستنده صحیح نیست"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StorePostalPackageRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "user1"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user1@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123123"
                    },
                    "image": {
                        "type": "string",
                        "example": "/users/images/user1.jpg"
                    }
                },
                "type": "object"
            },
            "PostalPackageResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "name": {
                        "type": "string",
                        "example": "user1"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user1@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-11 20:40:54"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-11 20:40:54"
                    },
                    "image": {
                        "type": "string",
                        "example": "/users/images/user1.jpg"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Postal Packages",
            "description": "Postal Packages"
        }
    ]
}